% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRepo.R
\name{createRepo}
\alias{createRepo}
\title{Função para criar um repositório padrão}
\usage{
createRepo(n)
}
\arguments{
\item{n}{defaults=empty}
}
\description{
Essa função cria um diretório padrão, com 8 pastas e um README.md
}
\examples{
createRepo()
}
\keyword{repository}
